update fair page and auth components
