+++ src/desktop/apps/artsy_in_miami/components/MiamiFairWeekPage.tsx
@@ -1,4 +1,3 @@
-import PropTypes from "prop-types"
 import React from "react"
 import styled, { ThemeProvider } from "styled-components"
 
@@ -23,8 +22,6 @@ const SectionTitle = styled(Title)`
 
 const IntroductionText = styled(Text)`
   line-height: 31px;
-  margin-bottom: 20px;
-  color: ${colors.grayDark};
   @media (max-width: 24em) {
     font-size: 20px;
     line-height: 26px;
@@ -36,15 +33,6 @@ const FairLogo = styled.img`
   display: inline;
   @media (min-width: 48em) {
     max-width: 160px;
-  }
-`
-
-const ResponsiveRow = styled(Row)`
-  ${props =>
-    props.paddingBottom &&
-    `padding-bottom: ${props.paddingBottom}px;`} @media (max-width: 48em) {
-    margin-left: -8px;
-    margin-right: -8px;
   }
 `
 
@@ -71,42 +59,53 @@ const theme = {
   },
 }
 
-export const FairWeekPageScaffold = ({
+interface MiamiFairWeekPageProps {
+  introduction: any
+  fair_coverage: any
+  artsy_in_miami: any
+  prepare_for_fairs: any
+}
+
+export const MiamiFairWeekPage: React.SFC<MiamiFairWeekPageProps> = ({
   introduction,
   fair_coverage,
-  event,
+  artsy_in_miami,
   prepare_for_fairs,
-  displayStickyFooter,
 }) => (
   <ThemeProvider theme={theme}>
     <Container>
-      <ResponsiveRow paddingBottom={50}>
+      <Row style={{ paddingBottom: 50 }}>
         <Col lg={4} md={4} sm={12} xs={12}>
           <SectionTitle
             titleSize="large"
             dangerouslySetInnerHTML={{ __html: introduction.title }}
           />
         </Col>
-
         <Col lg={8} md={8} sm={12} xs={12}>
           <ReveredColumnOnMobile>
-            <IntroductionText textSize="xlarge">
+            <IntroductionText
+              textSize="xlarge"
+              color={colors.grayDark}
+              style={{ marginBottom: 20 }}
+            >
               {introduction.description}
             </IntroductionText>
-            <img
-              style={{ marginTop: 30, marginBottom: 20, maxWidth: "100%" }}
-              src={introduction.image}
-            />
+            <div>
+              <img
+                style={{ marginTop: 30, marginBottom: 20, maxWidth: "100%" }}
+                src={introduction.image}
+              />
+            </div>
           </ReveredColumnOnMobile>
         </Col>
-      </ResponsiveRow>
+      </Row>
 
-      <ResponsiveRow paddingBottom={50}>
+      <Row style={{ paddingBottom: 50 }}>
         <Col lg={4} md={4} sm={12} xs={12}>
           <SectionTitle titleSize="large">{fair_coverage.title}</SectionTitle>
         </Col>
         <Col lg={8} md={8} sm={12} xs={12}>
-          <ResponsiveRow paddingBottom={20}>
+          <Row style={{ marginBottom: 20 }}>
             {fair_coverage.fairs.map(fair => (
               <Col lg={3} md={3} sm={3} xs={6} key={fair.logo_url}>
                 {fair.site_url && fair.site_url.startsWith("http") ? (
@@ -118,40 +117,38 @@ export const FairWeekPageScaffold = ({
                 )}
               </Col>
             ))}
-          </ResponsiveRow>
+          </Row>
         </Col>
-      </ResponsiveRow>
+      </Row>
 
-      {event && (
-        <ResponsiveRow paddingBottom={45}>
-          <Col lg={4} md={4} sm={12} xs={12}>
-            <SectionTitle titleSize="large">{event.title}</SectionTitle>
-          </Col>
-          <Col lg={8} md={8} sm={12} xs={12}>
-            <img
-              style={{ marginBottom: 10, width: "100%" }}
-              src={event.banner_image_url}
-            />
+      <Row style={{ paddingBottom: 45 }}>
+        <Col lg={4} md={4} sm={12} xs={12}>
+          <SectionTitle titleSize="large">{artsy_in_miami.title}</SectionTitle>
+        </Col>
+        <Col lg={8} md={8} sm={12} xs={12}>
+          <img
+            style={{ marginBottom: 10, width: "100%" }}
+            src={artsy_in_miami.banner_image_url}
+          />
 
-            <ResponsiveRow>
-              <Col lg={7} md={8} sm={12} xs={12} style={{ marginBottom: 25 }}>
-                <Text textSize="medium">{event.description}</Text>
-              </Col>
-              <Col lg={5} md={12} sm={12} xs={12} style={{ marginBottom: 25 }}>
-                <Text
-                  textSize="medium"
-                  color={colors.grayDark}
-                  dangerouslySetInnerHTML={{
-                    __html: event.public_viewing_date,
-                  }}
-                />
-              </Col>
-            </ResponsiveRow>
-          </Col>
-        </ResponsiveRow>
-      )}
+          <Row>
+            <Col lg={7} md={12} sm={12} xs={12} style={{ marginBottom: 25 }}>
+              <Text textSize="medium">{artsy_in_miami.description}</Text>
+            </Col>
+            <Col lg={5} md={12} sm={12} xs={12} style={{ marginBottom: 25 }}>
+              <Text
+                textSize="medium"
+                color={colors.grayDark}
+                dangerouslySetInnerHTML={{
+                  __html: artsy_in_miami.public_viewing_date,
+                }}
+              />
+            </Col>
+          </Row>
+        </Col>
+      </Row>
 
-      <ResponsiveRow>
+      <Row>
         <Col lg={4} md={4} sm={12} xs={12}>
           <SectionTitle titleSize="large">
             {prepare_for_fairs.title}
@@ -159,7 +156,7 @@ export const FairWeekPageScaffold = ({
         </Col>
         <Col lg={8} md={8} sm={12} xs={12}>
           {prepare_for_fairs.articles.map(article => (
-            <ResponsiveRow paddingBottom={25} key={article.title}>
+            <Row style={{ marginBottom: 25 }} key={article.title}>
               <Col lg={7} md={7} sm={6} xs={12}>
                 <a href={article.article_url} target="_blank">
                   <img
@@ -185,20 +182,10 @@ export const FairWeekPageScaffold = ({
                   </Text>
                 </a>
               </Col>
-            </ResponsiveRow>
+            </Row>
           ))}
         </Col>
-      </ResponsiveRow>
-
-      {displayStickyFooter && <div id="react-root-for-cta" />}
+      </Row>
     </Container>
   </ThemeProvider>
 )
-
-FairWeekPageScaffold.propTypes = {
-  introduction: PropTypes.object,
-  fair_coverage: PropTypes.object,
-  event: PropTypes.object,
-  prepare_for_fairs: PropTypes.object,
-  displayStickyFooter: PropTypes.bool,
-}
