+++ src/desktop/apps/artsy_in_miami/index.ts
@@ -1,19 +1,19 @@
-import { stitch as _stitch } from "@artsy/stitch"
+import queryString from "query-string"
+import { merge } from "lodash"
+import { stitch } from "@artsy/stitch"
+
 import adminOnly from "../../lib/admin_only"
 import JSONPage from "../../components/json_page/es6"
-import { FairWeekPageScaffold } from "desktop/components/fair_week_marketing/PageScaffold"
+import { MiamiFairWeekPage } from "./components/MiamiFairWeekPage"
 import { FairWeekMeta } from "desktop/components/fair_week_marketing/Meta"
-import { merge } from "lodash"
-import queryString from "query-string"
 
-let stitch = _stitch
-const SLUG = "spring-art-fairs"
-const MARKETING_MODAL_ID = "ca18"
+const SLUG = "artsy-in-miami"
+const MARKETING_MODAL_ID = "ca12"
 
-export class EditableFriezeWeekPage extends JSONPage {
+class EditableMiamFairWeekPage extends JSONPage {
   registerRoutes() {
     this.app.get(this.jsonPage.paths.show, this.show.bind(this))
-    this.app.get(this.jsonPage.paths.show + "/data", this.data)
+    this.app.get(this.jsonPage.paths.show + "/data", adminOnly, this.data)
     this.app.get(this.jsonPage.paths.edit, adminOnly, this.edit)
     this.app.post(this.jsonPage.paths.edit, adminOnly, this.upload)
   }
@@ -37,15 +37,11 @@ export class EditableFriezeWeekPage extends JSONPage {
         },
         blocks: {
           head: FairWeekMeta,
-          body: FairWeekPageScaffold,
-        },
-        locals: {
-          assetPackage: "banner_pop_up",
+          body: MiamiFairWeekPage,
         },
         data: {
           ...res.locals,
           ...data,
-          displayStickyFooter: !req.user,
           data,
         },
       })
@@ -57,7 +53,7 @@ export class EditableFriezeWeekPage extends JSONPage {
   }
 }
 
-export default new EditableFriezeWeekPage({
+export const app = new EditableMiamFairWeekPage({
   name: SLUG,
   paths: { show: `/${SLUG}`, edit: `/${SLUG}/edit` },
 }).app
