--- src/desktop/apps/artsy_in_miami/templates/meta.jade
@@ -1,5 +1,0 @@
-title= data.meta.title
-meta( property="og:title", content=data.meta.title )
-meta( name="description", content=data.meta.description )
-meta( name="og:description", content=data.meta.description )
-meta( property="twitter:description", content=data.meta.description )
