+++ src/desktop/apps/frieze_week/client/initializer.tsx
@@ -3,6 +3,7 @@ import ReactDOM from "react-dom"
 import { BannerPopUp } from "desktop/components/fair_week_marketing/BannerPopUp"
 
 export const init = () => {
+  // @ts-ignore
   const bootstrapData = window.__BOOTSTRAP__
   const {
     ctaTitle,
