+++ src/desktop/components/fair_layout/client/index.coffee
@@ -3,16 +3,15 @@ FairHeaderView = require './header.coffee'
 FooterView = require '../../main_layout/footer/view.coffee'
 lightDetector = require '../../light_detector/index.coffee'
 FairNavView = require './nav.coffee'
-MarketingSignupModal = require '../../marketing_signup_modal/index.coffee'
+{ triggerMarketingModal } = require '../../marketing_signup_modal/triggerMarketingModal.ts'
 
-module.exports = (options) =>
+module.exports = (options) ->
   globalClientSetup()
-
   { fair, model } = options
 
   new FairHeaderView el: $('.fair-layout-header'), model: model, fair: fair
   new FairNavView el: $('.fair-layout-nav'), model: model, fair: fair
-  new MarketingSignupModal
+  triggerMarketingModal(true)
 
   # check for feature image light/dark class
   if $('.feature-image').length
