+++ src/desktop/components/fair_week_marketing/Meta.tsx
@@ -1,20 +1,24 @@
-import PropTypes from "prop-types"
 import React from "react"
 
-export const FairWeekMeta = props => {
-  const { meta: { description, title } } = props
+interface FairWeekMetaProps {
+  meta: {
+    title: string
+    description: string
+  }
+}
+
+export const FairWeekMeta: React.SFC<FairWeekMetaProps> = props => {
+  const {
+    meta: { description, title },
+  } = props
 
   return (
-    <div>
+    <>
       <title>{title}</title>
       <meta property="og:title" content={title} />
       <meta property="description" content={description} />
       <meta property="og:description" content={description} />
       <meta property="twitter:description" content={description} />
-    </div>
+    </>
   )
-}
-
-FairWeekMeta.propTypes = {
-  meta: PropTypes.object,
 }
