--- src/desktop/components/fair_week_marketing/meta.jade
@@ -1,5 +1,0 @@
-title= meta.title
-meta( property="og:title", content=meta.title )
-meta( name="description", content=meta.description )
-meta( name="og:description", content=meta.description )
-meta( property="twitter:description", content=meta.description )
