+++ src/desktop/components/marketing_signup_modal/triggerMarketingModal.ts
@@ -1,0 +1,55 @@
+import { data as sd } from "sharify"
+import qs from "querystring"
+import { findWhere } from "underscore"
+const mediator = require("desktop/lib/mediator.coffee")
+
+export const triggerMarketingModal = (isScrolling?: boolean) => {
+  const query = qs.parse(location.search.replace(/^\?/, ""))
+  const isTargetCampaign = sd.CURRENT_PATH === sd.TARGET_CAMPAIGN_URL
+  const slug = query["m-id"] || (isTargetCampaign && "ca3")
+  const modalData = findWhere(sd.MARKETING_SIGNUP_MODALS, { slug: slug })
+
+  if (sd.MARKETING_SIGNUP_MODALS && modalData) {
+    if (!sd.CURRENT_USER && !sd.IS_MOBILE) {
+      if (isScrolling) {
+        scrollingMarketingModal(modalData)
+      } else {
+        staticMarketingModal(modalData)
+      }
+    }
+  }
+}
+
+export const scrollingMarketingModal = modalData => {
+  window.addEventListener(
+    "scroll",
+    () => {
+      setTimeout(() => {
+        const { image, copy } = modalData
+        mediator.trigger("open:auth", {
+          copy,
+          mode: "signup",
+          intent: "signup",
+          signupIntent: "signup",
+          trigger: "scroll",
+          triggerSeconds: 2,
+          destination: location.href,
+          image,
+        })
+      }, 2000)
+    },
+    { once: true }
+  )
+}
+
+export const staticMarketingModal = modalData => {
+  const { image, copy } = modalData
+  mediator.trigger("open:auth", {
+    copy,
+    mode: "signup",
+    intent: "signup",
+    signupIntent: "signup",
+    destination: location.href,
+    image,
+  })
+}
