+++ src/desktop/index.js
@@ -67,11 +67,11 @@ app.use(require("./apps/dev"))
 app.use(require("./apps/article").app)
 app.use(require("./apps/gallery_partnerships"))
 app.use(require("./apps/marketing_signup_modals"))
-app.use(require("./apps/artsy_in_miami").default)
-app.use(require("./apps/armory_week").default)
-app.use(require("./apps/frieze_week").default)
-app.use(require("./apps/frieze_week_london").default)
-app.use(require("./apps/basel_week").default)
+app.use(require("./apps/artsy_in_miami").app)
+app.use(require("./apps/armory_week").app)
+app.use(require("./apps/frieze_week").app)
+app.use(require("./apps/frieze_week_london").app)
+app.use(require("./apps/basel_week").app)
 
 // Non-profile dynamic vanity url apps
 app.use(require("./apps/galleries_institutions"))
